requirements.txt
whois
requests
beautifulsoup4
dnspython
user-agents
tldextract
shodan
reportlab
builtwith